modularise breadcrumb features
