+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/capture/crumbs/BreadcrumbDataSource.kt
