+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/injection/FeatureModuleImpl.kt
